Add files via upload
Added project folder containing all files to complete 1a. and 2a. The file app.java file that launches the SourceBuilder file 'DistributionChart.fxml' which contains the empty design for the bar chart. The java file 'mainframe.java' is the controller file for the fxml file. This file fills out all the data that is displayed on the fxml file.